requirements_teaser.txt
# Teaser Generator Dependencies
# Install with: pip install -r requirements_teaser.txt

# LLM & API Integration
openai>=1.3.0
requests>=2.31.0

# Environment Management
python-dotenv>=1.0.0

# Optional: Video processing (if using ffmpeg-python instead of subprocess)
# ffmpeg-python>=0.2.1

# Optional: Enhanced logging
# coloredlogs>=15.0

# Optional: Async support for parallel API calls
# aiohttp>=3.9.0
# asyncio-contextmanager>=1.0.0
